Allow multiple printPage iframe elements to print on page
We had an issue in production where we had 3 buttons, each of which would print a different url. After clicking on one button, each subsequent click would only print the first iframe in the document. The workaround we offered was to check if the element had an id, and append this to the id and name of the printPage element. This resolved our issues.